+++ Postfixform/Postfixform/Postfix.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <errno.h>
 
-int countTheExpression(char* postfixEntry)
+float countTheExpression(char* postfixEntry)
 {
     Stack* head = NULL;
     int counter = 0; 
@@ -11,7 +11,7 @@ int countTheExpression(char* postfixEntry)
     {
         if (postfixEntry[counter] >= '0' && postfixEntry[counter] <= '9')
         {
-            push(&head, postfixEntry[counter] - '0');
+            push(&head, (float)postfixEntry[counter] - '0');
             counter++;
             continue;
         }
@@ -20,12 +20,12 @@ int countTheExpression(char* postfixEntry)
             counter++;
             continue;
         }
-        int secondNumber = pop(&head);
+        float secondNumber = pop(&head);
         if (errno == 1)
         {
             return 0;
         }
-        int firstNumber = pop(&head);
+        float firstNumber = pop(&head);
         if (errno == 1)
         {
             return 0; 
@@ -54,7 +54,7 @@ int countTheExpression(char* postfixEntry)
         }     
         counter++;
     }
-    const int answer = pop(&head);
+    const float answer = pop(&head);
     if (errno == 1)
     {
         return 0;
